integrate with gqlgen
