+++ services/accounts/graph/accounts.go
@@ -1,0 +1,16 @@
+package graph
+
+import "github.com/ContextLogic/accounts/graph/model"
+
+var users = []*model.User{
+	{
+		ID:       "1",
+		Name:     "Ada Lovelace",
+		Username: "@ada",
+	},
+	{
+		ID:       "1",
+		Name:     "Alan Turing",
+		Username: "@complete",
+	},
+}
