+++ services/inventory/gqlgen.yml
@@ -54,3 +54,9 @@ models:
       - github.com/99designs/gqlgen/graphql.Int
       - github.com/99designs/gqlgen/graphql.Int64
       - github.com/99designs/gqlgen/graphql.Int32
+  Product:
+    fields:
+      inStock:
+        resolver: true
+      shippingEstimate:
+        resolver: true
