+++ services/products/graph/entity.resolvers.go
@@ -5,14 +5,11 @@ package graph
 
 import (
 	"context"
-	"fmt"
-
 	"github.com/ContextLogic/products/graph/generated"
 	"github.com/ContextLogic/products/graph/model"
 )
 
 func (r *entityResolver) FindProductByUpc(ctx context.Context, upc string) (*model.Product, error) {
-	fmt.Printf("FindProductByUpc: %s\n", upc)
 	for _, product := range products {
 		if product.Upc == upc {
 			return product, nil
@@ -25,34 +22,3 @@ func (r *entityResolver) FindProductByUpc(ctx context.Context, upc string) (*mod
 func (r *Resolver) Entity() generated.EntityResolver { return &entityResolver{r} }
 
 type entityResolver struct{ *Resolver }
-
-// !!! WARNING !!!
-// The code below was going to be deleted when updating resolvers. It has been copied here so you have
-// one last chance to move it out of harms way if you want. There are two reasons this happens:
-//  - When renaming or deleting a resolver the old code will be put in here. You can safely delete
-//    it when you're done.
-//  - You have helper methods in this file. Move them out to keep these resolver files clean.
-var products []*model.Product
-
-func init() {
-	products = []*model.Product{
-		{
-			Upc:    "1",
-			Name:   "Table",
-			Price:  899,
-			Weight: 100,
-		},
-		{
-			Upc:    "2",
-			Name:   "Couch",
-			Price:  1299,
-			Weight: 1000,
-		},
-		{
-			Upc:    "3",
-			Name:   "Chair",
-			Price:  54,
-			Weight: 50,
-		},
-	}
-}
