+++ services/products/graph/products.go
@@ -1,0 +1,26 @@
+package graph
+
+import (
+	"github.com/ContextLogic/products/graph/model"
+)
+
+var products = []*model.Product{
+	{
+		Upc:    "1",
+		Name:   "Table",
+		Price:  899,
+		Weight: 100,
+	},
+	{
+		Upc:    "2",
+		Name:   "Couch",
+		Price:  1299,
+		Weight: 1000,
+	},
+	{
+		Upc:    "3",
+		Name:   "Chair",
+		Price:  54,
+		Weight: 50,
+	},
+}
