+++ services/reviews/gqlgen.yml
@@ -54,3 +54,11 @@ models:
       - github.com/99designs/gqlgen/graphql.Int
       - github.com/99designs/gqlgen/graphql.Int64
       - github.com/99designs/gqlgen/graphql.Int32
+  User:
+    fields:
+      reviews:
+        resolver: true
+  Product:
+    fields:
+      reviews:
+        resolver: true
